requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-environ==0.11.2
requests==2.32.3
django-cors-headers==4.4.0

